박민재 sprint4

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -5,6 +5,8 @@ import com.sprint.mission.discodeit.dto.request.MessageUpdateRequest;
 import com.sprint.mission.discodeit.dto.response.MessageUpdateResponse;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.Message;
+import com.sprint.mission.discodeit.exception.FileException;
+import com.sprint.mission.discodeit.exception.ObjectNotFoundException;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.MessageRepository;
@@ -13,7 +15,9 @@ import com.sprint.mission.discodeit.service.MessageService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -31,25 +35,30 @@ public class BasicMessageService implements MessageService {
 
     //interface
     @Override
-    public Message createMessage(MessageCreateRequest request) {
-        //입력값 검증 처리하겠습니다
-        validateString(request.content());
-        validateUUID(request.authorId());
-        validateUUID(request.channelId());
-
+    public Message createMessage(MessageCreateRequest request, List<MultipartFile> files) {
         //존재하는 유저, 채널인지 검증
         validateUserExists(request.authorId());
         validateChannelExists(request.channelId());
 
-        //첨부파일 추가 작업
         List<UUID> binaryContentIdList = new ArrayList<>();
-        if (request.attachmentPathList() != null) {
-            for (String attachmentPath : request.attachmentPathList()) {
-                if (attachmentPath != null && !attachmentPath.isBlank()) {
-                    //binaryContent 생성
-                    BinaryContent binaryContent = new BinaryContent(attachmentPath);
-                    binaryContentRepository.createBinaryContent(binaryContent);
-                    binaryContentIdList.add(binaryContent.getId());
+        //첨부파일 존재 시
+        if (files != null) {
+            for (MultipartFile file : files) {
+                if (file != null && !file.isEmpty()) {
+                    try {
+                        //binaryContent 생성
+                        BinaryContent binaryContent = new BinaryContent(
+                                file.getOriginalFilename(),
+                                (long) file.getBytes().length,
+                                file.getContentType(),
+                                file.getBytes()
+                        );
+                        binaryContentRepository.createBinaryContent(binaryContent);
+                        binaryContentIdList.add(binaryContent.getId());
+
+                    } catch (IOException e) {
+                        throw new FileException(e.getMessage());
+                    }
                 }
             }
         }
@@ -64,38 +73,45 @@ public class BasicMessageService implements MessageService {
 
     @Override
     public List<Message> findAllByChannelId(UUID channelId) {
-        //입력값 검증 처리하겠습니다
-        validateUUID(channelId);
-
         return messageRepository.findAllMessagesByChannelId(channelId);
     }
 
     @Override
-    public MessageUpdateResponse updateMessage(MessageUpdateRequest request) {
-        //입력값 검증 처리하겠습니다
-        validateUUID(request.messageId());
-        validateString(request.content());
-
+    public MessageUpdateResponse updateMessage(MessageUpdateRequest request, List<MultipartFile> files) {
         //메시지 검색
         Message messageTemp = messageRepository.findMessageById(request.messageId())
-                .orElseThrow(() -> new RuntimeException("에러: 해당 메시지는 데이터파일에 존재하지 않습니다."));
+                .orElseThrow(() -> new ObjectNotFoundException("에러: 해당 메시지는 데이터파일에 존재하지 않습니다."));
 
-        //이전 첨부파일 삭제
-        for (UUID attachmentId : messageTemp.getAttachmentIds()) {
-            binaryContentRepository.deleteBinaryContent(attachmentId);
-        }
-
-        //첨부파일 추가 작업
-        List<UUID> binaryContentIdList = new ArrayList<>();
-        if (request.attachmentPathList() != null) {
-            for (String attachmentPath : request.attachmentPathList()) {
-                if (attachmentPath != null && !attachmentPath.isBlank()) {
-                    //binaryContent 생성
-                    BinaryContent binaryContent = new BinaryContent(attachmentPath);
-                    binaryContentRepository.createBinaryContent(binaryContent);
-                    binaryContentIdList.add(binaryContent.getId());
+        List<UUID> binaryContentIdList = messageTemp.getAttachmentIds();
+        //첨부파일 존재 시
+        if (files != null) {
+            //이전 첨부파일 삭제
+            if (binaryContentIdList != null) {
+                for (UUID attachmentId : binaryContentIdList) {
+                    binaryContentRepository.deleteBinaryContent(attachmentId);
                 }
             }
+
+            List<UUID> newBinaryContentIdList = new ArrayList<>();
+            for (MultipartFile file : files) {
+                if (file != null && !file.isEmpty()) {
+                    try {
+                        //binaryContent 생성
+                        BinaryContent binaryContent = new BinaryContent(
+                                file.getOriginalFilename(),
+                                (long) file.getBytes().length,
+                                file.getContentType(),
+                                file.getBytes()
+                        );
+                        binaryContentRepository.createBinaryContent(binaryContent);
+                        newBinaryContentIdList.add(binaryContent.getId());
+
+                    } catch (IOException e) {
+                        throw new FileException(e.getMessage());
+                    }
+                }
+            }
+            binaryContentIdList = newBinaryContentIdList;
         }
 
         log.info("메시지: {}가 수정됨.\n->{}", messageTemp.getContent(), request.content());
@@ -109,12 +125,9 @@ public class BasicMessageService implements MessageService {
 
     @Override
     public void deleteMessage(UUID messageId) {
-        //입력값 검증 처리하겠습니다
-        validateUUID(messageId);
-
         //메시지 검색
         Message messageTemp = messageRepository.findMessageById(messageId)
-                .orElseThrow(() -> new RuntimeException("에러: 해당 메시지는 데이터파일에 존재하지 않습니다."));
+                .orElseThrow(() -> new ObjectNotFoundException("에러: 해당 메시지는 데이터파일에 존재하지 않습니다."));
 
         //첨부파일 삭제
         for (UUID attachmentId : messageTemp.getAttachmentIds()) {
@@ -128,28 +141,16 @@ public class BasicMessageService implements MessageService {
     }
 
 
-    // 들어온 String 필드가 null 혹은 공백인지 검증하는 메서드
-    private void validateString(String str) {
-        if (str == null || str.isBlank()) {
-            throw new IllegalArgumentException("에러: 입력값이 Null 또는 공백입니다.");
-        }
-    }
-    // 들어온 UUID 필드가 null인지 검증하는 메서드
-    private void validateUUID(UUID id) {
-        if (id == null) {
-            throw new IllegalArgumentException("에러: 입력값이 Null입니다.");
-        }
-    }
     // 들어온 userId 필드가 레포지터리에 존재하는지 검증하는 메서드
     private void validateUserExists(UUID userId) {
         if (!userRepository.existsUserById(userId)) {
-            throw new RuntimeException("유저: " + userId + "이 존재하지 않습니다.");
+            throw new ObjectNotFoundException("유저: " + userId + "이 존재하지 않습니다.");
         }
     }
     // 들어온 channelId필드가 레포지터리에 존재하는지 검증하는 메서드
     private void validateChannelExists(UUID channelId) {
         if (!channelRepository.existsChannelById(channelId)) {
-            throw new RuntimeException("채널: " + channelId + "이 존재하지 않습니다.");
+            throw new ObjectNotFoundException("채널: " + channelId + "이 존재하지 않습니다.");
         }
     }
 }
